+++ packages/atlas/src/views/viewer/MembershipSettingsView/ChangePasswordDialog.tsx
@@ -1,72 +1,19 @@
 import styled from '@emotion/styled'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { u8aToHex } from '@polkadot/util'
-import axios from 'axios'
 import { FC, useCallback, useState } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 
-import { SvgActionHide, SvgActionShow } from '@/assets/icons'
 import { PasswordCriterias } from '@/components/_auth/PasswordCriterias'
 import { FormField } from '@/components/_inputs/FormField'
-import { Input, InputProps } from '@/components/_inputs/Input'
+import { Input } from '@/components/_inputs/Input'
 import { DialogModal } from '@/components/_overlays/DialogModal'
-import { atlasConfig } from '@/config'
-import { ORION_AUTH_URL } from '@/config/env'
-import { keyring } from '@/joystream-lib/lib'
-import { decodeSessionEncodedSeedToMnemonic, prepareEncryptionArtifacts } from '@/providers/auth/auth.helpers'
+import { useHidePasswordInInput } from '@/hooks/useHidePasswordInInput'
+import { changePassword } from '@/providers/auth/auth.helpers'
 import { useAuth } from '@/providers/auth/auth.hooks'
 import { useSnackbar } from '@/providers/snackbars'
 import { media, sizes } from '@/styles'
 import { passwordAndRepeatPasswordSchema } from '@/utils/formValidationOptions'
 import { SentryLogger } from '@/utils/logs'
-
-type ChangePasswordArgs = {
-  joystreamAccountId: string
-  gatewayAccountId: string
-  email: string
-  encodedSeed: string
-  newPassword: string
-}
-export const changePassword = async ({
-  email,
-  encodedSeed,
-  newPassword,
-  joystreamAccountId,
-  gatewayAccountId,
-}: ChangePasswordArgs) => {
-  try {
-    const mnemonic = await decodeSessionEncodedSeedToMnemonic(encodedSeed)
-    if (!mnemonic) {
-      throw Error(`Couldn't get mnemonic`)
-    }
-
-    const timestamp = Date.now()
-    const keypair = keyring.addFromMnemonic(mnemonic)
-    const newArtifacts = await prepareEncryptionArtifacts(email, newPassword, mnemonic)
-
-    const changePasswordPayload = {
-      joystreamAccountId,
-      gatewayName: atlasConfig.general.appName,
-      timestamp,
-      action: 'changeAccount',
-      gatewayAccountId,
-      newArtifacts,
-    }
-
-    const signatureOverPayload = u8aToHex(keypair.sign(JSON.stringify(changePasswordPayload)))
-
-    return axios.post(
-      `${ORION_AUTH_URL}/change-account`,
-      {
-        signature: signatureOverPayload,
-        payload: changePasswordPayload,
-      },
-      { withCredentials: true }
-    )
-  } catch (error) {
-    SentryLogger.error('Something went wrong during changing password', 'changePassword', error)
-  }
-}
 
 type ChangePasswordDialogProps = {
   onClose: () => void
@@ -78,8 +25,6 @@ type PasswordStepForm = {
   confirmPassword: string
 }
 
-type PasswordInputNames = keyof PasswordStepForm
-
 export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, show }) => {
   const { displaySnackbar } = useSnackbar()
 
@@ -87,11 +32,8 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
     shouldFocusError: true,
     resolver: zodResolver(passwordAndRepeatPasswordSchema),
   })
-
-  const [isFieldVisible, setIsFieldVisible] = useState<Record<PasswordInputNames, boolean>>({
-    password: false,
-    confirmPassword: false,
-  })
+  const hidePasswordProps = useHidePasswordInInput()
+  const hideConfirmPasswordProps = useHidePasswordInInput()
 
   const handleClose = () => {
     form.reset({ password: '', confirmPassword: '' })
@@ -104,7 +46,7 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
   const handleChangePassword = useCallback(() => {
     form.handleSubmit(async (data) => {
       if (!currentUser || !encodedSeed) {
-        return
+        throw Error('Current user or enodedSeed is not set')
       }
       try {
         setIsSubmiting(true)
@@ -132,25 +74,6 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
     })()
   }, [currentUser, displaySnackbar, encodedSeed, form, onClose])
 
-  const handleTogglePassword = (name: PasswordInputNames) => {
-    setIsFieldVisible((fields) => ({ ...fields, [name]: !fields[name] }))
-  }
-  const getInputProps = useCallback(
-    (name: PasswordInputNames): InputProps => {
-      return {
-        ...form.register(name),
-        type: isFieldVisible[name] ? 'text' : 'password',
-        actionButton: {
-          children: isFieldVisible[name] ? 'Hide' : 'Show',
-          dontFocusOnClick: true,
-          icon: isFieldVisible[name] ? <SvgActionHide /> : <SvgActionShow />,
-          onClick: () => handleTogglePassword(name),
-        },
-      }
-    },
-    [form, isFieldVisible]
-  )
-
   return (
     <StyledDialogModal
       show={show}
@@ -168,10 +91,15 @@ export const ChangePasswordDialog: FC<ChangePasswordDialogProps> = ({ onClose, s
     >
       <Wrapper>
         <FormField label="New password" error={form.formState.errors.password?.message}>
-          <Input placeholder="New password" {...getInputProps('password')} autoComplete="off" />
+          <Input placeholder="New password" {...form.register('password')} {...hidePasswordProps} autoComplete="off" />
         </FormField>
         <FormField label="Repeat Password" error={form.formState.errors.confirmPassword?.message}>
-          <Input placeholder="Repeat password" {...getInputProps('confirmPassword')} autoComplete="off" />
+          <Input
+            placeholder="Repeat password"
+            {...form.register('confirmPassword')}
+            {...hideConfirmPasswordProps}
+            autoComplete="off"
+          />
         </FormField>
         <FormProvider {...form}>
           <PasswordCriterias />
